Provide a working 404 by using markdown
In order for the 404 to be picked up by jekyll/github pages,
it has to be in a 404.html file on the root. However, if the
file is named exactly '404.html', it won't be processed by
jekyll.

Rather, you have to provide a 404.md file, and specify its
permalink to be 404.html, which fixes the issue.

Working version at http://www.cazzulino.com/wrong-url.html.